Changed image color overlay approach
Instead of setting background image to an empty div, I put the two images (for mobile and desktop) in the HTML markup and then changed their display properties depending on the media query. I also created a div with class after, to which I gave background color (the "accent" color) and blended it with the image.